requirements.txt
python-dotenv==0.19.1
python-telegram-bot==13.7
google-cloud-dialogflow==2.9.1
